Add Id column to suites Restructure suites components Refactoring
